Corrected inequality operator from <> to /=.
Changed "use declarations" to "use mean_param", and added wind_n as an
argument.  This is a reversion to how the code was in kcollinsTC and
dlatornell06 versions.  It resolves compilation errors that arise due
to repeated declarations of year, day, and wind arguments.

Did lots of code clean-up and applied emacs f90-indent-subprogram
function.  Added CVS Id and Source keywords to top of file.