Corrected format of data with the command: DougsTools/fixDataFile.py "%i %i %i %i %i %.1f %.1f %.1f %.1f %.1f %.1f %.1f %.1f %.1f %.1f %.1f %.1f %.1f %.1f %.1f %.1f %.1f %.1f %.1f %.1f %.1f %.1f %.1f %.1f" <input/met/yvr/hum200123456.dat >input/met/yvr/hum200123456.dat.fixed
This reformats the file to put stn, year, month, day, and para data
(1st 5 fields) in integer representation instead of e-format, and air
humidity data (remaining 24 fields) in 1 decimal place real
format. This is necessary because when the code is compiled with g95
it is stricter than pgf90 in its handling of free-format input.