Corrected format of data with the command: DougsTools/fixDataFile.py "%i %i %i %e %e" <input/rivers/eng200123456.dat >input/rivers/eng200123456.dat.fixed
This reformats the file to put year, month, and day data (1st 3
fields) in integer representation instead of e-format, and river flow
and mystery data (remaining 2 fields) in e-format format. This is
necessary because when the code is compiled with g95 it is stricter
than pgf90 in its handling of free-format input.