Initial commit.
Bottom CTD data file that include 2006 data, but is missing from
kcollins_NH4 in repository.

This file needs to be checked for inconsistencies like Susan found in
the met data, but I'll use it for comparison with kcollins_NH4 code.
For production runs it will be replaced by a file in sog-forcing/.