Significant re-factoring:
Wrapped subroutine in module so that we get compile-time checking of
argument number and types.

Deleted USE of surface_forcing because it's not required.

Changed order of arguments to put outputs (unow & vnow) last.

Changed pretty much all of the local variable names to make them more
self-explanatory.  Deleted deltat local because it wasn't really
necessary.

Deleted write statements in loop that searches wind data structure
because they were really quasi-deprecated debugging code that clutters
stdout.

Changed the variables and values that are output with the "search
failed" error message so that they should be more useful.

Added lots of documentation, and suggestions for future improvements
(latter delimited by *** at beginnning of comment).