Added use precision_defs.
Changed the setting of the Coriolos factor from a value to calculate
it from latitude.  Latitude is still set as a constant, but there are
notes for improvement suggesting that latitude should be read as a run
parameter, and f calculated from it during the run initialization.

This produced a small change in the value of f, which causes the
results of the nov04-test run to not diff cleanly against reference
results, but the profiles of salinity and nitrate show no visible
change.