Changes resulting from the refactoring of Cp_mod.f90 into water_properties module (and a few unrelated cleanups).
Added use water_properties, grid_mod, and defind_flux_mod modules.

Declared Cp as type water_property.

Added call to alloc_water_props to allocated memory for Cp.

Removed Cp from coefficients call.

Moved intialization of water column properties to before the beginning
of the time loop; eliminating the if (time_step == 1 .and. count == 1)
check in the iteration loop.

Added Cp%i as an argument to define_flux.

Replaced 2 calls of Cp_sub with call to Cp_profile, and interp_i
functions.

Deleted declaration of variables dS, Sa, and smooth_i that were never
set and never used.

Deleted declaration and initialization of bin_tot2 variable that was
set but never used.

The change in Cp formulation results in minor changes to the profiles
and time series output so that they no longer diff cleanly against the
previous reference copies.  The differences are not significant on
profile plots, although minor differences remain in the profiles
generated by runs on coho and on iBooks.