Eliminated find_PON as it was only diagnostic
Use P_H for both particles and nutrients -- this eliminates the subroutine
N_H.  The only difference in the formation of the Hvector is the sinking term.
Use the null vector for the nutrients.

Set the sinking term for P%micro explicitly to zero.

added calculation of Silicon Hvector and solution by TRIDAG solver.

This set of modification has no effect on the results.