fitbottom is a module containing the bottom boundary conditions.  There are currently 3 subroutines.  The first initializes the constants.  This one should be changed to parameter definitions --> Doug!
The second one uses a seasonal fit, for all the quantities we have data
for, to set the bottom boundary condition.  This is a change from previously
which used a linear change from one CTD cast to the next.  For most of the
variables, this means we are following noise in the data.  Note that
salinity does seem to have a year over year trend!!!

The third subroutine uses a zero gradient boundary condition for the
quantities we have no data for.