Reworked the derivs_sog code to use PZ_bins.  Changed variable names so they make more sense.  Put only's on the use statements.
1. Corrected error in previous derivs_sog that used PZ instead of Yplus to set
NO and NH.  (My oops)

2. Added waste%small to NH and detritus derivatives.

3. Added a check for negative NH.

The first substantive change makes a significant difference when running with
flagellates.

An error in the formulation of the basic NH/Detritus equations means that
flagellates make a small difference.  This version will not diff clean
against previous non-flagellate runs.

(Error will be corrected next... but it again will give a different answer
and not the same as the original).

This version can be run with or without flagellates.