Added a missing space in the *FieldUnits header line of the profiles results file format statement.
Added the halocline magnitude and depth to the header of the profiles
results file.

Changed the name suffix of the profiles results file to use no
separator between the time components (i.e. 2004-10-29T120700 instead
of 2004-10-29T12:07:00) because : plays havoc with file name
completion in most unix shells, and appears also to trigger a bug in
Python.  I'm betting we can blame this all on Bill Gates...