Order of items has been re-organized and comments added to make best use of new input_processor.
Parameters values for grid and run timing that were formerly read from
the input/*.dat file labelled here as inputfile are now contained in
this file.  The input file lable has been removed, and corresponding
*.dat files can be deleted.

The comments in input/biology_sog.dat and input/nutrient.dat
concerning the provenance of various parameter values for the biology
model could be copied to the bottom of this file as notes, and those
files deleted.

Please update any infiles that you have created based on this template
to reflect these changes.