Replaced declarations of getpar* functions with a "use input_processor, only: ..." statement.
Deleted last argument in all getpar* function invocations because the
report argument is now optional, defaulting to true, so only required
when you want to suppress reporting of the parameter value to stdout.

Added CTDdatetime as an argument to init_timeseries_output(), and
write_timeseries(), and added *InitialCTDDateTime: field containing
the CTDdatetime value to the headers of all of the time series results
files.