Added a new private read_grid() subroutine to read the grid parameter values from the infile  read_grid() is called by init_grid().
Changed the grid spacing parameter (lambda) from a public variable to
a private one because it is now longer required outside of the module.

Changed the grid interface and point index arrays into pointers to
arrays, and allocated them in alloc_grid() because their size is now
longer known when init_grid is called.