Replaced declarations of getpar* functions with a "use input_processor, only: ..." statement.
Deleted last argument in all getpar* function invocations because the
report argument is now optional, defaulting to true, so only required
when you want to suppress reporting of the parameter value to stdout.

Added startDatetime as an argument to init_profiles_output(), and
write_profiles(), so that a *InitialCTDDateTime: field containing the
startDatetime value appears in the headers of all of the profiles, and
time series results files.

Moved the write_timeseries() call from the end of the time step loop
to just after the end of the implicit solver loop, with the
write_profiles() call.  This shifts the time series results by 1 time
step - new reference files will need to be saved.

Reorganized the initialization section of the code to put contents of
infile in more logical order (IMHO).  infiles derived from the
template need to be updated.