New module to handle the forcing data including the wind, river and met-data (does not do the bottom boundary conditions... that is in fitbottom)
Three public subroutines and a number of service subroutines.

read_forcing : reads the forcing data from the files and stores them in arrays
internal to forcing.f90

get_forcing : given year, day and time returns the values of the wind, rivers
and met_data.

read_variation : reads from the inputfile requests to shift the year of the
rivers, wind or cloud fraction, change the amount by a fraction or set equal
to a fix value.  This allows the varaiations that Kate did in her thesis/paper
without manipulating the data files (only changing the input files)