Deleted unused argument hh.
Deleted clouds from arguments list because the values that it brought
in are now assigned within this subroutine.

Moved cloudy, and okta type definitions in from mean_params.f90
because they are really local to this subroutine.

Moved assignment of cloud%type array values in from coefficients.f90
because they are local to this subroutine, and coefficients.f90 is
going away.