Added use statements for the core_variables, and physics_model modules.
Added calls to alloc_core_variables(), and dalloc_core_variables() to
handle their memory allocation.

Changed T to T%new in the initial_mean() call because the latter is
the only thing that initial_mean() sets.

Added a gradient_i() function call at the top of the implicit solver
loop to calculate the temperature gradient profile at the grid
interface depths (T%grad_i).  This replaces calls to div_interface()
farther down, and in the double_diffusion() subroutine.

T%grad_i replaces T%div_i so that the gradient_i() service function in
the grid module can be used for both core variables, and water
property variables.

Replaced the call to double_diff() with a call to the refactored
double_diffusion() subroutine in the physics_model module.  Note that
the order of arguments has been standardized, and T%grad_i is passed
in instead of the whole T array structure.

Added T%grad_i to argument list in define_flux() call.