Finished refactoring of double_diffusion() subroutine.
Replaced grid in argument list with number of grid points, because
that is all that is needed in the subroutine.

Replaced sal in the argument list with S_grad_i because S%grad_i is
the only component of S that is used in the subroutine.

Expanded the nus argument to nu_t_dd, and nu_s_dd to make the
relationship to K%t%dd, and K%s%dd in SOG.f90 more obvious.

Deleted use statements for grid_mod, mean_param, and surface_forcing
because the subroutine now uses nothing from any of those modules.

Added more comments about what the code is doing, and references to
equations in Large, et al (1994).