Added Coriolis factor (f) and pi as public parameter values.
Coriolis factor is not actually, strictly speaking, a parameter
because it has to be calculated from the latitude in init_physics() in
order to keep the pgf90 compiler happy.  But routines that use if
shouldn't have to worry about that.

There is commented out code for both g and f values that provides more
accurate values.  That code will be implemented shortly.