Changed tridiag() calls for Detritus arrays so that results are written into %n32 components instead of temporary arrays.
Refactored code for dealing with negative values in Detritus arrays
out of find_new.f90 into SOG (temporarily), and added printing of
warning messages when negative values arise.  Also moved setting of
Detritus surface boundary conditions.  This eliminates the need for
the find_new() subroutine, so its call was removed, and it can be
removed from the project.