Corrected comment.
Previous commit by me (so 2 commits ago) as well as mortality changes
I corrected an error.

Under light limitation with plenty of NH the NO uptake was being set
to zero instead of its previous value.  (So if the flagellates have
enough NH but the diatoms didn't the NO uptake was being wrongly set
to zero).

This correction causes the results to not diff cleanly.