Changes related to refactoring of arrays for RHS terms of semi-implicit diffusion/advection equations for biological quantities.
Created a new subroutine call new_diffusion_ceoff() to load diffusion
coefficients into vectors of tridiagonal matrix instead of using a
derived type variable for the tridiagonal matrix.  This is a temporary
subroutine because we need to keep the original diffusion_coeff()
around until the physics equations RHS terms are similarily
refactored.

Renamed the Gvector argument of the diffusion_bot_surf_flux()
subroutine to RHS to reflect the change of Gvector to RHS arrays.