Changed argument list of the build_biology_equations() subroutine to remove the diff_coeffs%*, and *_RHS% arrays.  Having them both used in SOG.f90 and passed as arguments created an effective EQUIVALENCE that is probably deprecated programming practice.
Added comments to alert readers of the source code to the fact that
the subroutine operates on public variables not in its argument list.

Renamed diff_coeffs%* to diff_coeffs_bio% for future compatibility
with diffusion coefficient matrices for physics quantities.

The diff_coeffs_bio%*, and *_RHS%* variables are public and used in
SOG.f90 as a temporary measure during refactoring.  They will
eventually only be used by other modules (probably just the yet to be
created IMEX_solver module) and not used in SOG.f90.