Changes related to renaming and restructuring of biology modules and subroutines.
Added day and time arguments to solve_bio_eqns() subroutine so that
they can be passed through to the check_negative() subroutine.

Added a new check_bio_negatives() subroutine as a wrapper around calls
to check_negative() for each of the biology quantities.  This replaces
the negative checking code that was in SOG.f90, having been refactored
out of the old find_new() subroutine.