Updated the sinking algorithm including:
P_micro_sink is now an array and the sinking_advection algorithm takes an
array.  To facilitate sinking for constant sinkers (detritus PON and BSi) a
unit vector (1,1,1...) is defined.

Sinking algorithm completely changed as Nicoles very complicated algorithm
was only valid for constant w or other non-divergent flows.  Sinking is not
necessarily non-divergent.  Now using upwind schemce.

Also changed wupwell to start at 0 as it is now defined on the interfaces.