Changes related to physics equations refactoring:
Deleted the call to define_sog().  It has been replaced by
new_to_old_*() subroutines in various modules, and removed from the
project.

Deleted calls to all of the diffusion_*()subroutines for physics
quantities because they have been refactored into the
physics_eqn_builder module.

Reworked the physics equations refactoring bridge code to work around
the upwelling and Coriolis_and_pg subroutines.  The former can't be
refactored until Susan finishes some work on the freshwater, and
find_upwell modules.  The latter is awaiting the movement of the
baroclinic pressure gradient calculation code out of the physics_model
module into is own module, to resolve circular use conditions.

Replaces all of the new to old copy code for Bmatrix*%*, and
Gvector*%* with calls to new_to_old_*() subroutines.

Replaced all of the calls to matrix_A(), phys_Hvector(), and tridiag()
with a call to the new solve_phys_eqns() subroutine in the IMEX_solver
module.