Changed use statement that bring the value of f in from physics_model to fundamental_constants.
Removed use statement for dPdx_b, and dPdy_b becuase they are now used
behind the scenes, and don't need to pass through SOG.

Added call to init_contants() subroutine to set the value of the
Coriolis factor (f), all because of a pgf90 bug that prevents it from
being set as a parameter due to its use of the sin() intrinsic.

Deleted use of the Coriolis_and_pg_mod module because that code is now
part of the physics_eqn_builder module.  An interesting side effect
was that the value of the precision_defs sp parameter was being
supplied to SOG by Coriolis_and_pg_mod.  Corrected that by adding sp
to the use precision_defs statement.

Deleted the calls to the Coriolis_and_pg() subroutine and the
corresponding physics equations refactoring bridge code.