Added type definitions, and variable declarations for new diffusion profile arrays: nu%*, K_ML%*, and K%*, the interior, mixing layer, and overall diffusion values, respectively.
Added type definitions, and variable declarations for new turbulent
kinematic fluxes profile arrays: wbar%*.  *** Note that only the
surface values (index = 0) are used in the model, but that code to
calculate the full profiles is available in ***.  It is commented out,
but can be included in a build if it is required for diagnostic
purposes.

Added memory allocation/deallocation subroutines, and code in them for
diffusion, and turbulent kinematic fluxes profile arrays.