Changes related to turbulence refactoring:
Deleted double_diffusion() subroutine from use physics_model statement
because that subroutine is now in the turbulence module where it is
private.

Moved call to double_diffusion(), and the code that sets the values of
the total interior diffusion coefficients (K%*%total) into
calc_KPP_diffusivity().  Added refactoring bridge code so that the
rest of the not yet refactored turbulence code can deal with these
changes.