Added a new lower bound (lb) argument to the interp_value() subroutine.  Doing so makes it more robust, and eliminates the silliness of sometimes having to subtract 1 from the j_below value it returns.
Adjusted the lower bound and size of the indices array to support the
lower bound argument addition to interp_value().

Moved the alloc_grid() & dalloc_grid() subroutines to the end of the
module for consistency with other modules.

Added the lower bound argument to all of the interp_value() calls in
the depth_average() subroutine.  Eliminated the aforementioned
subtract 1 silliness.  Did whitespace cleanup.

Changed the out of range condition in interp_value() from a warning to
an error message on stderr that stops execution.