A number of things:
1) Changed the F_n flux salinity calculation to be a constant salinity
(the deep water value) times Fw.  This is because the estuarine circulation
doesn't really dump water on the surface but rather advects a gradient in
salinity by S3.  The salinity scale of this gradient is the differnce of the
deep water salinity to the river salinity (0).  The strength of the advection
is proportional to the river flow and hence to Fw.

2) Removed Bf blending.  No longer necessary as the freshwater flux is now
a pretty stable forcing.

3) Carefully considered the convergence scales for velocity and mixed layer
depth and they appear directly in SOG.  Thus I removed del_o

4) Average both the mixed layer depth (as we were doing previously) and the
velocity.  This greatly reduces the number of iterations so I could somewhat
increase the precision (through the scales above).

5) Added a variation in the averaging to "try something different" if it is not
converging.  This solves about 1/2 of the non-convergence cases.  Currently
nov04-test is running through with 2 non-convergences.

(6) Tried changing the minimum number of iterations.... absolutely no evidence
this improves anything.