Implemented new maximum temperature for growth of phytoplankton so that, for example Thalassiosira will not grow at 18 or more degrees.
Implemented new light scheme for phytoplankton based on the Steeves curve but
with extend growth at high light.  This matches the Durbin curve for T.n.
nicely and also matches a slightly extended version of the S.c. curve.

These changes change the biological parameters that are read into the model.