Replaced all stop statements with call exit(1) statements.  This causes the Fortran runtime to return an exit status of 256 to the operating system.  That enables us to detect in scripts when the code has terminated abnormally.
Note that call exit() is an extension that is present in both pgf90
and g95, but is not guaranteed to be there in all Fortran compilers,
so this change could cause trouble if the code is ever ported to
another compiler.

Changed all warning and error messages to write to stdout instead of
stderr.  It seems that messages written to stderr (unit 0) get lost
when the output of the run is being redirected.

Added a call exit(0) statement to the end of the program so that we
explicitly return an exit status of 0 to the operating system when the
run completes normally.