Reorganized the use statements at the top of the file, and added comments to the use statements for all of the variables that are imported from other modules.
Deleted a bunch of local variables that hae been refactored into the
numerics module.

Replaced the code that read the loop time parameters from the infile,
and did some other initialization with a call to the new
init_numerics() subroutine in the numerics module.  Code for the
implicit solver loop convergence stabilization variables, and
convergence metrics has also been refactored into numerics.

Moved the call to the init_grid() subroutine to before the call to
init_numerics() because we can't do much until we know the size of the
grid.

Renamed the maximum iteration limit variable from niter to max_iter.

Streamlined the code that blends values of h, U, & V from the current
and previous iteration.

Cleaned up the convergence metrics code.