Added use statements for the new buoyancy module and its variables.
Changed all instances of call buoyancy() to call calc_buoyancy() to
reflect the name change in the new buoyancy module.  Bf is no longer
an argument of calc_buoyancy(), so added comments to draw attention to
the fact that it is a public variable that is set by the subroutine.