Added a new public type called timedelta to hold date/time differences.  They are expressed in days and seconds.
Moved leap year determination code into its own function called
leapday() that returns 0 or 1 depending on whether the year given as
its argument contains a leap day or not.

Modernized the defintion of the day_sec() function.  Lots more scope
for code cleanup in this module.

Added a new function called datetime_incr() to increment a datetime
variable by a timedelta variable.  Only positive increments are
allowed, and the largest allowable is 365 days and 86399 seconds.