Renamed public startDatetime variable to initDatatime to emphasize the fact that it is the date/time of the initial conditions CTD profile used to initialize the run.
Added a new public datetime type variable, curDatetime to hold the
date/time of the current time step through the run.  This will replace
the year, day & day_time variables, and be incremented at the end of
each time step.

Added 2 new timedelta type variables, dur & ndt.  dur is private to
the module and used to calculate the duration of the run.  ndt will
eventually replace the public real variable dt.

Replaced reading of initial conditions date/time components with a
getpar_datetime() call that reads the initial date/time from a
date/time string.

Replaced reading of run end time (t_f) with a getpar_datetime() call
that reads the end date/time from a date/time string.

Changed reading of time step (dt) to read an integer into the new ndt
variable.  It is immediately cast to dble() in dt for now.

Rewrote the code that calculates the number of time steps in terms of
the new variables.