Use init_core_variables from core_variables module in place of both alloc_core_variables, and initial_mean from initial_sog.
***NOTE: this change alters the order of items in the infile.  All infiles will have to be updated.***
read_variation() and init_core_variables() are now called just after init_numerics() so the variation and initialization data files sections of infile move from the end to near the top.

This is part of the refactoring of the core_variables module to include initialization of the core variable profiles.